requirements.txt
requests==2.32.5
tiktoken==0.8.0
sentence-transformers==3.3.1
supabase==2.10.0
python-dotenv==1.0.1
